Windows/MSVC portability: POSIX shim refactor, drop root include/
- Fold the MSVC POSIX compatibility shim into src/mbio/mb_define.h (block at
  end, under #ifdef _WIN32); force-included on every TU via /FI mb_define.h.
- Add src/mbio/unistd_w.h (<unistd.h> substitute) and select it per-site and
  in mb_define.h: #ifdef _WIN32 -> unistd_w.h, #else -> <unistd.h>.
- Add src/mbio/dirent_w.h (minimal Win32 opendir/readdir/closedir) for
  mbmakedatalist; guard its <dirent.h> the same way.
- Remove root include/ (mb_win_compat.h, unistd.h); put src/mbio on the MSVC
  include path.
- Guard src/surf/xdr_win32.h XdrOp/XDR with _MB_WIN_XDR_DEFINED so the
  force-included mb_define.h copy does not collide.
- Build mbfnv2navlab, mbmakedatalist, mbnavlab2fnv.
- Sync remaining non-TRN port work (gsf_enc.c, src/gmt, src/mr1pr, .def files).
- Reset TRN sources (src/mbtrn, mbtrnav, mbtrnframe, mbtrnutils) to original.

+++ src/mbtrnav/trnw/mcsub.c
@@ -404,8 +404,8 @@ int main(int argc, char *argv[])
 
         	// format and print message
             // summary
-            fprintf(stderr,"SUB - mrx msg[%-*s] len[%7lld]  src[%s : %hu]\n",
-                    wmsg, (aout_en?rxbuf:""), (long long)tx_bytes,
+            fprintf(stderr,"SUB - mrx msg[%-*s] len[%7lu]  src[%s : %hu]\n",
+                    wmsg, (aout_en?rxbuf:""), tx_bytes,
                     inet_ntoa(x_addr.sin_addr),
                     htons(x_addr.sin_port));
 
@@ -435,8 +435,8 @@ int main(int argc, char *argv[])
             if( (tx_bytes=sendto(fd, txbuf, tx_len, 0, (struct sockaddr *) &x_addr,
                                addrlen))>0){
 
-                fprintf(stderr,"SUB - utx msg[%-*s] len[%3zu/%-3lld] dest[%s : %hu]\n",
-                        wmsg, txbuf, (size_t)tx_len,(long long)tx_bytes,
+                fprintf(stderr,"SUB - utx msg[%-*s] len[%3lu/%-3lu] dest[%s : %hu]\n",
+                        wmsg, txbuf, tx_len,tx_bytes,
                         inet_ntoa(x_addr.sin_addr),
                         htons(x_addr.sin_port));
 
@@ -449,8 +449,8 @@ int main(int argc, char *argv[])
                 if ((rx_bytes=recvfrom(fd, rxbuf, MSGBUFSIZE, 0,
                                      (struct sockaddr *) &x_addr, &addrlen)) > 0){
 
-                    fprintf(stderr,"SUB - urx msg[%-*s] len[%7lld]  src[%s : %hu]\n",
-                            wmsg, rxbuf, (long long)rx_bytes,
+                    fprintf(stderr,"SUB - urx msg[%-*s] len[%7lu]  src[%s : %hu]\n",
+                            wmsg, rxbuf, rx_bytes,
                             inet_ntoa(x_addr.sin_addr),
                             htons(x_addr.sin_port));
                 }
